Actualizacion Igna 1.0.1

+++ vite-project/src/components/Card/Card.jsx
@@ -2,34 +2,39 @@
 //react
 import { useState, useEffect } from "react";
 import { useNavigate } from "react-router-dom";
-import { useDispatch, useSelector } from "react-redux";
-// import { PostFavoriteHotel, DeleteFavoriteHotel } from "../../redux/Actions/Actions.js";
-import { Link } from "react-router-dom";
+import { PedirLocalStorage } from "../Index";
+import { useSelector, useDispatch } from "react-redux";
+import { FuncionAllFavoritesHotel } from "../../redux/Actions/Actions.js";
 import style from "./Card.module.css";
-
 import axios from "axios";
 import imgFav from "../../image/favorito.png";
 
 //?----------------- COMPONENTE CARD ------------------------------------
 function Cards({ id, name, image, province, rating, description, valoration }) {
+  const dispatch = useDispatch();
   const navigate = useNavigate();
-  const idUser = useSelector((state) => state.idUser);
   const FavHotels = useSelector((state) => state.FavHotels);
   const [isFav, setIsFav] = useState(false);
   const URL_BASE = "https://las-casitas-del-hornero-back-deploy.up.railway.app";
   let ratingArray = Array(rating).fill(rating);
 
+  let User = PedirLocalStorage();
+
   useEffect(() => {
     FavHotels?.forEach((fav) => {
-      if (fav.id == id) setIsFav(true);
+      if (fav.id === id) setIsFav(true);
     });
   }, [FavHotels]);
 
   const handleFavorite = async (idUser, id) => {
-    setIsFav(!isFav);
-    isFav
-      ? await axios.delete(`${URL_BASE}/favorites/${idUser}/${id}`)
-      : await axios.post(`${URL_BASE}/favorites/${idUser}/${id}`);
+    console.log(idUser);
+    if (idUser) {
+      setIsFav(!isFav);
+      isFav
+        ? await axios.delete(`${URL_BASE}/favorites/${idUser}/${id}`)
+        : await axios.post(`${URL_BASE}/favorites/${idUser}/${id}`);
+      dispatch(FuncionAllFavoritesHotel(User.id));
+    }
   };
 
   const onClickDetail = () => {
@@ -42,14 +47,14 @@ function Cards({ id, name, image, province, rating, description, valoration }) {
         <img src={image[0]} alt="" className={style.imgHotel} />
         {isFav ? (
           <button
-            onClick={() => handleFavorite(idUser, id)}
+            onClick={() => handleFavorite(User?.id, id)}
             className={style.button}
           >
             ❤️
           </button>
         ) : (
           <button
-            onClick={() => handleFavorite(idUser, id)}
+            onClick={() => handleFavorite(User?.id, id)}
             className={style.button}
           >
             🤍
@@ -59,7 +64,7 @@ function Cards({ id, name, image, province, rating, description, valoration }) {
       <div className={style.title}>
         <h4 className={style.name}>{name}</h4>
         <p>{valoration}</p>
-        <h7 className={style.province}>{province}</h7>
+        <h3 className={style.province}>{province}</h3>
         <div className={style.rating}>
           {ratingArray.map((_, index) => {
             return (
